requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.0
tzdata==2025.2
Werkzeug==3.1.3
tensorflow==2.15.0

# py -3.11 -m venv .venv
# pip install -r requirements.txt